Combine LanceFragment.merge_columns and LanceDataset.add_columns into a single method
